# [5.2] Improve performance of validating arrays