## Benchmark

#### Data:

```php
$data = [
	"contacts" => [
		[
			"id" => 1,
			"name" => "Contact Name",
			"phones" => [
				[
					"number" => "(###) ###-####",
					"type" => "other"
				], // ...
			],
			"addresses" => [
				[
					"address" => "address ...",
					"type" => "home"
				], // ...
			],
		], // ...
	]
];
```

#### Rules:

```php
$v = \Validator::make($data, [
	'contacts.*.id' => 'required|integer|min:1',
	'contacts.*.name' => 'required|string|between:3,50',
	'contacts.*.phones.*.number' => 'required|string|regex:/^\(\d{3}\) \d{3}-\d{4}$/',
	'contacts.*.phones.*.type' => 'required|string|in:home,work,other',
	'contacts.*.addresses.*.type' => 'required|string|in:home,work,other',
	'contacts.*.addresses.*.address' => 'required|string|between:10,300',
]);
```

#### Results:

*Note: opcache is enabled (fpm + cli)*

- **10 contacts, 5 phones and 5 addresses:**

| PHP 5.6 (Before)   | PHP 5.6 (After)   | PHP 7.0 (Before)  | PHP 7.0 (After)   | 
| ------------------:|:-----------------:|:-----------------:|:-----------------:|
| 0.040791988372803  | 0.033037900924683 | 0.014898061752319 | 0.011796951293945 |
| 0.041944980621338  | 0.032547950744629 | 0.011984825134277 | 0.012818098068237 |
| 0.041891098022461  | 0.033796072006226 | 0.012062788009644 | 0.011888027191162 |
| 0.040940999984741  | 0.035719871520996 | 0.012029170989990 | 0.011542797088623 |
| 0.044970989227295  | 0.034888029098511 | 0.012119054794312 | 0.011468887329102 |
| 0.037815093994141  | 0.032665014266968 | 0.012550830841064 | 0.011487960815430 |

- **50 contacts, 5 phones and 5 addresses:**

| PHP 5.6 (Before)   | PHP 5.6 (After)   | PHP 7.0 (Before)  | PHP 7.0 (After)   | 
| ------------------:|:-----------------:|:-----------------:|:-----------------:|
| 0.16860485076904   | 0.13122105598450  | 0.039514064788818 | 0.035233020782471 |
| 0.17176103591919   | 0.12970399856567  | 0.040139913558960 | 0.035001993179321 |
| 0.16790485382080   | 0.13179397583008  | 0.042943000793457 | 0.035140037536621 |
| 0.16598796844482   | 0.13261985778809  | 0.039919853210449 | 0.035305976867676 |
| 0.16733813285828   | 0.13014101982117  | 0.039392948150635 | 0.036513090133667 |
| 0.16732406616211   | 0.13396883010864  | 0.047128915786743 | 0.035287857055664 |

- **200 contacts, 10 phones and 10 addresses:**

| PHP 5.6 (Before)   | PHP 5.6 (After)   | PHP 7.0 (Before)  | PHP 7.0 (After)   | 
| ------------------:|:-----------------:|:-----------------:|:-----------------:|
| 1.8555521965027    | 1.0786969661713   | 0.35779786109924  | 0.25464081764221  |
| 1.8819801807404    | 1.0602869987488   | 0.36769986152649  | 0.25297498703003  |
| 1.8798329830170    | 1.0668759346008   | 0.36662507057190  | 0.26263403892517  |
| 1.8572149276733    | 1.0674540996552   | 0.36115002632141  | 0.25368118286133  |
| 1.8599259853363    | 1.0681769847870   | 0.36576199531555  | 0.26442599296570  |
| 1.9706079959869    | 1.0951921939850   | 0.36076998710632  | 0.25122284889221  |

### Conclusion:

#### PHP 7.0:

| Data                                          | Before (Average)  | After (Average)   | Improvement     | 
|----------------------------------------------:|:-----------------:|:-----------------:|:---------------:|
| **10 contacts, 5 phones and 5 addresses:**    | 0.012607455253601 | 0.011833786964416 | 6.5% faster     |
| **50 contacts, 5 phones and 5 addresses:**    | 0.04150644938151  | 0.03541366259257  | 17.2% faster    |
| **200 contacts, 10 phones and 10 addresses:** | 0.36330080032348  | 0.25659664471944  | 41.5% faster    |

#### PHP 5.6:

| Data                                          | Before (Average)  | After (Average)   | Improvement     | 
|----------------------------------------------:|:-----------------:|:-----------------:|:---------------:|
| **10 contacts, 5 phones and 5 addresses:**    | 0.04139252503713  | 0.033775806427002 | 22.5% faster    |
| **50 contacts, 5 phones and 5 addresses:**    | 0.16815348466237  | 0.13157478968303  | 27.8% faster    |
| **200 contacts, 10 phones and 10 addresses:** | 1.8841857115428   | 1.072780529658    | 75.6% faster    |

*Also PHP 7.0 is almost 4x faster than PHP 5.6 in this case! WOW!*