@taylorotwell 2015-12-28
I assume all these results are in milliseconds?