@GrahamCampbell 2015-12-28
How does the peak memory usage compare?