@alibo 2015-12-28
@taylorotwell No, they aren't. All of them are in seconds (`microtime(true)`).

@GrahamCampbell Some results seem weird specially for PHP 7.0!

#### Memory Usage:

**PHP 7.0:**

| Data                           | Before - Usage/Total (kB) | After - Usage/Total (kB) | Increase        | 
|----------------------------------------------:|:-----------------:|:-----------------:|:---------------:|
| **10 contacts, 5 phones and 5 addresses:**    | 0/6144            | 0/6144            | 0% - 0kB       |
| **50 contacts, 5 phones and 5 addresses:**    | 0/6144            | 0/6144            | 0% - 0kB       |
| **200 contacts, 10 phones and 10 addresses:** | 6144/14336        | 6144/14336        | 0% - 0kB       |

**PHP 5.6:**

| Data                           | Before - Usage/Total (kB) | After - Usage/Total (kB) | Increase        | 
|----------------------------------------------:|:-----------------:|:-----------------:|:---------------:|
| **10 contacts, 5 phones and 5 addresses:**    | 256/8192          | 256/8192          | 0% - 0kB       |
| **50 contacts, 5 phones and 5 addresses:**    | 1024/9216         | 1280/9472         | 25% - 256kB    |
| **200 contacts, 10 phones and 10 addresses:** | 9216/20224        | 10240/21248       | 11.1% - 1024kB |

And this is the code:

~~~php
$before = memory_get_peak_usage(true);

$v = \Validator::make($data, [/* rules */]);
$v->passes();

$after = memory_get_peak_usage(true);

$this->info('Total Memory Usage: ' . ($after / 1024) . ' KB') ;
$this->info('Validating Memory Usage: ' . (($after - $before) / 1024) . ' KB');
~~~