@alibo 2015-12-28
@taylorotwell 
> If get_as_float is set to TRUE, then microtime() returns a float, which represents the current time in seconds since the Unix epoch accurate to the nearest microsecond.

http://php.net/manual/en/function.microtime.php