@GrahamCampbell 2015-12-28
Ha, the worst thing for me is the inconstancy between `int` and `integer`, and simelar things.