@alibo 2015-12-28
@GrahamCampbell ha ha it's funny!! :D also this one:
https://3v4l.org/e4PIu#v700rc4
> Return value of add() must be an instance of **integer**, **integer** returned

We're changing the topic :))