@GrahamCampbell 2015-12-28
> Return value of add() must be an instance of integer, integer returned

Haha. That's hilarious. That surely has to be a bug!